fix(kernels): round takes half inputs; the campaign's video gate compares frames
round refused half-precision inputs like floor/ceil (libdevice nearbyint is
fp32/fp64 only): every --triton run of the Kokoro speech model stopped at
its first aten::round. A half value is rounded in fp32 (exact, half to
even) and stored back in its dtype; fp32 unchanged. round joins the bank
(0 ULP on every shape; 198 references, 53 of 280 kernels reached).

tools/precision_zoo_campaign.py: the video gate decodes both clips and
compares frames (per-frame PSNR, the image gate's 30 dB bar, sha kept) —
a byte-only gate cannot judge a lever on a video; `regate` re-runs the
gate on existing outputs without the GPU. Re-gated: Allegro's two arms are
11.8 dB apart (8.7 dB worst frame) — and both render smeared colour bands
at the campaign's stimulus, so the row is an invalid reference, not a
lever verdict (filed).

+++ src/neurobrix/kernels/ops/floor.py
@@ -36,7 +36,8 @@ def round_forward_kernel(input_ptr, output_ptr, n_elements, BLOCK_SIZE: tl.const
     # tl.math.nearbyint does not exist in Triton 3.6; libdevice.nearbyint is the
     # round-half-to-even primitive matching torch.round (cf. pow.py using
     # tl.extra.cuda.libdevice.pow). The round kernel was crashing before this.
-    tl.store(output_ptr + offset, tl.extra.cuda.libdevice.nearbyint(x), mask=mask)
+    # A half input is rounded in fp32 (exact) and stored back in its dtype (Kokoro --triton, 2026-09-05).
+    tl.store(output_ptr + offset, tl.extra.cuda.libdevice.nearbyint(x.to(tl.float32)).to(x.dtype), mask=mask)
 
 
 @triton.jit
